Actually move SearchFiltersPanel to kolibri-common
