+++ kolibri/core/assets/src/core-app/apiSpec.js
@@ -43,8 +43,6 @@ import commonTaskStrings from '../mixins/taskStrings';
 import commonSyncElements from '../mixins/commonSyncElements';
 import translatedUserKinds from '../mixins/userKinds';
 import CoreFullscreen from '../views/CoreFullscreen';
-import SearchFiltersPanel from '../views/SearchFiltersPanel';
-import SidePanelModal from '../views/SidePanelModal';
 import * as exams from '../exams/utils';
 import * as validators from '../validators';
 import * as objectSpecs from '../objectSpecs';
@@ -188,8 +186,6 @@ export default {
       MasteryModel,
       NotificationsRoot,
       KolibriLoadingSnippet,
-      SidePanelModal,
-      SearchFiltersPanel,
     },
     componentSets: {
       sync,
