+++ kolibri/plugins/learn/assets/src/views/TopicsPage/TopicsPanelModal.vue
@@ -50,7 +50,7 @@
 
   import useKResponsiveWindow from 'kolibri-design-system/lib/composables/useKResponsiveWindow';
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
-  import SidePanelModal from 'kolibri.coreVue.components.SidePanelModal';
+  import SidePanelModal from 'kolibri-common/components/SidePanelModal';
   import commonLearnStrings from '../commonLearnStrings';
   import useContentLink from '../../composables/useContentLink';
