+++ packages/kolibri-common/components/SearchBox.vue
