+++ packages/kolibri-common/components/SearchFiltersPanel/ActivityButtonsGroup.vue
