+++ packages/kolibri-common/components/SearchFiltersPanel/CategorySearchModal/CategorySearchOptions.vue
