+++ packages/kolibri-common/components/SearchFiltersPanel/CategorySearchModal/index.vue
