+++ packages/kolibri-common/components/SearchFiltersPanel/SelectGroup.vue
