+++ packages/kolibri-common/components/SearchFiltersPanel/index.vue
